+++ server/client/valentin/src/dashboard/compontents/health/health.detail.js
