+++ server/client/valentin/src/dashboard/compontents/health/health.detail.test.js
@@ -10,7 +10,7 @@ global.IntersectionObserver = IntersectionObserver;
   
 import React from 'react';
 import { render } from '@testing-library/react';
-import DescDetail from './healthDetail';
+import DescDetail from './health.detail.js';
 
 // Eine einfache Komponente, die das Test-Icon rendert
 const TestIconComponent = () => {
