+++ server/client/valentin/src/dashboard/compontents/logstream/logstream.item.js
@@ -1,0 +1,17 @@
+
+import {  Code, Tr, Td } from "@chakra-ui/react"
+
+function LogstreamItem({ type, schema, message }) {
+    
+  return (
+
+    <Tr>
+        <Td>
+            <Code colorScheme={schema}>{type}</Code>
+        </Td>
+        <Td>{message}</Td>
+    </Tr>
+  )
+}
+
+export default LogstreamItem;
