+++ server/client/valentin/src/dashboard/dashboard.js
@@ -1,36 +1,32 @@
 
 import { ChakraProvider } from "@chakra-ui/react"
-import { Button } from "@chakra-ui/react"
-import { Container, Box, Center } from '@chakra-ui/react'
+import { Box, Center } from '@chakra-ui/react'
 import { Fade } from "react-awesome-reveal";
 
-import Health from './compontents/health/healthBase';
+import Health from './compontents/health/health.base';
 import Head from './compontents/head';
-import GrafanaValidation from "./compontents/grafana/grafana.validation";
+// import GrafanaValidation from "./compontents/grafana/grafana.validation.js";
+import Logstream from "./compontents/logstream/logstream.js";
 
 function Dashboard({ children }) {
   return <ChakraProvider>{children}
-        <Center>
-        
-        
-        
-        
-        <Box m={5} width="100%" maxWidth={'1400px'}>
-          <Head></Head>    
-        
-          <Fade>
-            <Health/>
-          </Fade>
+              <Center>  
+                <Box m={5} width="100%" maxWidth={'1400px'}>
+                  <Head></Head>    
+                
+                  <Fade>
+                    <Health/>
+                  </Fade>
 
-          <Fade>
-           <GrafanaValidation/>
-          </Fade>
+                  <Fade>
+                    <Logstream/>
+                  </Fade>
 
-
-        </Box>
-        </Center>
-
-
+                  {/* <Fade>
+                  <GrafanaValidation/>
+                  </Fade> */}
+                </Box>
+              </Center>
     </ChakraProvider>
 }
